fix(ETISSSystemCWrapper): correct memory read/write on 32b system
Earlier the ETISS was trying to read/write memory using 64b e.g.
data from 0x80000000 was written erroneously on 0xffffffff80000000.
This caused MMU exceptions for 32b systems.

In this commit, we detect if the ISS is to be employed in a 32b
setting and mask the ETISS addresses appropriately on the
SystemC wrapper before sending out TLM transactions outside.