fix(elfLoader): elf file now properly loaded via cli in ETISS mode
Previously we used a hard-coded path for proof-of-concept which was
ugly and relied on recompiling the simulator in case target-sw
binary path changed.

Now we refactored code to forward the already used CLI option
`--elf` to `system.h/cpp` module. Further, we use better
exception-handling to remove ambiguity in case of wrong elf paths.